* Makefile.in: made to work under Windows    - added windows only make preamble    - moved the all target after the objects it depends on    - used $(OBJEXT) and $(EXEEXT) where needed    - split crctest into unix rules and windows rules    - changed file separatars into macros where needed * crctest.c: changed int to long to work on 16 bit machines * crc.c: added Windows calling convention to the function    and casted the assignment of the function. [There's a minor problem -- the cast should be fixed another way --  but I'll check it in this way and then Keith will fix it.]
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@4949 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/crypto/crc32/crctest.c
@@ -38,7 +38,7 @@ main()
     char input[16], expected_crc[16];
     unsigned char inbytes[4], outbytes[4];
     int in_length;
-    unsigned int expect;
+    unsigned long expect;
 
     int bad = 0;
